+++ src/note_conversion.c
@@ -4,7 +4,7 @@
 
 // --------
 
-int Convert_musical_to_pitch(char symbol, int half, int* pitch) {
+int Convert_musical_to_pitch(char symbol, int8_t half, int8_t* pitch) {
 
 	if(pitch == NULL) {
 		return 1;
